(chore) rename TermHistory->TermReviewHistory; (type) use actual TermHistory type in TermReviewHistory
